[#110] refactor: customRepository 위치 변경
