+++ src/main/java/goorm/eagle7/stelligence/domain/vote/custom/DummyVoteRepository.java
@@ -1,6 +1,8 @@
-package goorm.eagle7.stelligence.domain.vote;
+package goorm.eagle7.stelligence.domain.vote.custom;
 
 import org.springframework.stereotype.Component;
+
+import goorm.eagle7.stelligence.domain.vote.VoteSummary;
 
 /**
  * 빈을 찾지 못해 에러가 발생하는 것을 방지하기 위해 임시 빈을 등록합니다.
